requirements.txt
Babel==2.6.0
Django==2.1.7
django-phonenumber-field==2.2.0
djangorestframework==3.9.1
phonenumbers==8.10.6
Pillow==5.4.1
psycopg2-binary==2.7.7
pytz==2018.9
